Create SOAP_Si.py
This code is used to encodes the changes of the amorphous Silicon Structure after MD simulation. The SOAP Fingerprint of the starting structure is compared with 3 Silicon phases- LDA-Si, HDA-Si and cubic diamond Silicon.
Similar comparison can be done with the last frame of the MD simulation to identify the changes in the local atomic environments.